Cleanup covenant descriptor module
